requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
jinja2==3.1.4
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
httpx==0.27.0
requests==2.32.3